Fix Position on Qt 6.4

+++ src/application/position.hpp
@@ -17,7 +17,6 @@
 #define POSITION_HPP
 
 #include <QObject>
-#include <cstdint>
 #include <sstream>
 
 namespace noteahead {
@@ -26,31 +25,31 @@ namespace noteahead {
 struct Position
 {
     Q_GADGET
-    Q_PROPERTY(size_t pattern MEMBER pattern)
-    Q_PROPERTY(size_t track MEMBER track)
-    Q_PROPERTY(size_t column MEMBER column)
-    Q_PROPERTY(size_t line MEMBER line)
-    Q_PROPERTY(size_t lineColumn MEMBER lineColumn)
+    Q_PROPERTY(quint64 pattern MEMBER pattern)
+    Q_PROPERTY(quint64 track MEMBER track)
+    Q_PROPERTY(quint64 column MEMBER column)
+    Q_PROPERTY(quint64 line MEMBER line)
+    Q_PROPERTY(quint64 lineColumn MEMBER lineColumn)
 
 public:
     //! Index of the pattern.
-    size_t pattern = 0;
+    quint64 pattern = 0;
 
     //! Index of the active track of the pattern.
-    size_t track = 0;
+    quint64 track = 0;
 
     //! Index of the active note column of the track.
-    size_t column = 0;
+    quint64 column = 0;
 
     //! Index of the active line (common for all tracks).
-    size_t line = 0;
+    quint64 line = 0;
 
     //! Index of the active column of the line consisting of note and velocity values.
     //! 0: Note, e.g. "C-3"
     //! 1: Velocity digit 2, e.g. [1]27
     //! 2: Velocity digit 1, e.g. 1[2]7
     //! 3: Velocity digit 0, e.g. 12[7]
-    size_t lineColumn = 0;
+    quint64 lineColumn = 0;
 
     bool operator==(const Position & other) const
     {
